# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.CreateSettingsRequest:
    properties:
      description:
        type: string
      group:
        type: string
      is_public:
        type: boolean
      label:
        type: string
      setting_key:
        type: string
      type:
        type: string
      value_bool:
        type: boolean
      value_float:
        type: number
      value_int:
        type: integer
      value_string:
        type: string
    type: object
...
  models.SettingsResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      group:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      label:
        type: string
      setting_key:
        type: string
      type:
        type: string
      updated_at:
        type: string
      value_bool:
        type: boolean
      value_float:
        type: number
      value_int:
        type: integer
      value_string:
        type: string
    type: object
  models.SettingsSelectOption:
    properties:
      id:
        type: integer
      name:
        description: Display name
        type: string
    type: object
...
  models.UpdateSettingsRequest:
    properties:
      description:
        type: string
      group:
        type: string
      is_public:
        type: boolean
      label:
        type: string
      setting_key:
        type: string
      type:
        type: string
      value_bool:
        type: boolean
      value_float:
        type: number
      value_int:
        type: integer
      value_string:
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  types.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/types.Pagination'
    type: object
  types.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  types.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: info@negenet.com
    name: Negenet Team
    url: https://negenet.com
  description: This is the API documentation for Negenet
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://negenet.com/terms
  title: Negenet API
  version: 2.0.0
paths:
...
  /settings:
    get:
      consumes:
      - application/json
      description: Get a list of settings
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Sort field (id, created_at, updated_at,setting_key,label,group,type,value_string,value_int,value_float,value_bool,description,is_public,)
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List settings
      tags:
      - App/Settings
    post:
      consumes:
      - application/json
      description: Create a new Settings with the input payload
      parameters:
      - description: Create Settings request
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/models.CreateSettingsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create a new Settings
      tags:
      - App/Settings
  /settings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Settings by its id
      parameters:
      - description: Settings id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete a Settings
      tags:
      - App/Settings
    get:
      consumes:
      - application/json
      description: Get a Settings by its id
      parameters:
      - description: Settings id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a Settings
      tags:
      - App/Settings
    put:
      consumes:
      - application/json
      description: Update a Settings by its id
      parameters:
      - description: Settings id
        in: path
        name: id
        required: true
        type: integer
      - description: Update Settings request
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update a Settings
      tags:
      - App/Settings
  /settings/all:
    get:
      consumes:
      - application/json
      description: Get a simplified list of all settings with id and name only (for
        dropdowns/select boxes)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SettingsSelectOption'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List all settings for select options
      tags:
      - App/Settings
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: Enter your token with the prefix "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"